Broke things up some
